data-processing/nats/telemetry-stream/telemetry-stream-subscriber/pyproject.toml
[project]
name = "telemetry-stream-subscriber"
version = "1.0.0"
requires-python = "==3.14.3"
dependencies = [
  "nats-py==2.12.0",
  "pycapnp==2.2.2",
]

[dependency-groups]
dev = [
  "capnp-stub-gen==0.1.3",
  "poethepoet==0.41.0",
  "pytest-cov==7.0.0",
  "pytest==9.0.2",
]

[tool.uv]
package = false
required-version = ">=0.10.0"

[tool.poe.tasks]
generate-types = "capnp-stub-gen batch ../schemas/telemetry.capnp --output=.generated/capnp_types"
dev = { cmd = "python src/main.py", env = { PYTHONPATH = ".generated" }, deps = [
  "generate-types",
] }
test = { cmd = "pytest --verbose --verbose", env = { PYTHONPATH = ".generated" }, deps = [
  "generate-types",
] }
test-coverage = { cmd = "pytest --cov=. --cov-report=xml", env = { PYTHONPATH = ".generated" }, deps = [
  "generate-types",
] }
